comments crud and posts crud added...minor changes required after common service jar is imported

+++ post-microservice/src/main/java/com/microservice/post/service/PostService.java
@@ -1,11 +1,12 @@
 package com.microservice.post.service;
 
 import com.microservice.post.reponse.Response;
-import com.microservice.post.repository.PostRepository;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 
+@Slf4j
 @Service
 public class PostService {
 
@@ -13,25 +14,78 @@ public class PostService {
     private PostRepository postRepository;
 
     public Response getPosts() {
-        return new Response("Post Returned from database", Boolean.TRUE, "",  postRepository.findAll());
+        try {
+            log.info("Post Returned from database");
+            return new Response("Post Returned from database", Boolean.TRUE, "", postRepository.findAll());
+        } catch (Exception e) {
+            log.error("Post could not be returned from database due to exception... " + e.getMessage());
+            return new Response("Post could not be returned from database due to exception", Boolean.FALSE, "", null);
+        }
     }
 
+
     public Response addPost(Post post) {
-        postRepository.save(post);
-        return new Response("Post Added Successfully", Boolean.TRUE, "", null);
+        try {
+            postRepository.save(post);
+            log.info("Post Added Successfully");
+            return new Response("Post Added Successfully", Boolean.TRUE, "", null);
+        } catch (Exception e) {
+            log.info("Post could not be added due to exception... " + e.getMessage());
+            return new Response("Post could not be added due to some exception", Boolean.FALSE, "", null);
+        }
+
     }
 
     public Response getPostById(int id) {
-        return new Response("Post Returned from database", Boolean.TRUE, "",  postRepository.findById(id));
+        try {
+            if (postRepository.existsById(id)) {
+                log.info("Post Returned from database");
+                return new Response("Post Returned from database", Boolean.TRUE, "", postRepository.findById(id));
+            } else {
+                log.info("Post with id: " + id + " does not exist in database");
+                return new Response("Post with id: " + id + " does not exist in database", Boolean.FALSE, "", null);
+            }
+        } catch (Exception e) {
+            log.info("Post could not be returned due to some exception... " + e.getMessage());
+            return new Response("Post could not be returned due to some exception", Boolean.FALSE, "", null);
+        }
+
     }
 
     public Response updatePost(int id, Post post) {
-        postRepository.save(post);
-        return new Response("Post Updated", Boolean.TRUE, "", null);
+        try {
+            if (postRepository.existsById(id)) {
+                Post post1 = postRepository.findById(id);
+                // post1 = post
+                postRepository.save(post);
+                log.info("Post Updated Successfully");
+                return new Response("Post Updated Successfully", Boolean.TRUE, "", null);
+            } else {
+                log.info("Post with id: " + id + " does not exist in database");
+                return new Response("Post with id: " + id + " does not exist in database", Boolean.FALSE, "", null);
+            }
+        } catch (Exception e) {
+            log.info("Post could not be updated due to some exception... " + e.getMessage());
+            return new Response("Post could not be updated due to some exception", Boolean.FALSE, "", null);
+        }
+
     }
 
     public Response deletePostById(int id) {
-        postRepository.deleteById(id);
-        return new Response("Post Deleted Successfully", Boolean.TRUE, "", null);
+        try {
+            if (postRepository.existsById(id)) {
+                postRepository.deleteById(id);
+                log.info("Post Deleted Successfully");
+                return new Response("Post Deleted Successfully", Boolean.TRUE, "", null);
+            } else {
+                log.info("Post with id: " + id + " does not exist in database");
+                return new Response("Post with id: " + id + " does not exist in database", Boolean.FALSE, "", null);
+            }
+        } catch (Exception e) {
+            log.info("Post could not be deleted due to some exception... " + e.getMessage());
+            return new Response("Post could not be deleted due to some exception", Boolean.FALSE, "", null);
+        }
+
 
     }
+}
